refactor: 단일 서비스 클래스 구조로 전환
- 팀 컨벤션 변경
